requirements.txt
annotated-types==0.7.0
appdirs==1.4.4
attrs==25.3.0
blinker==1.7.0
build==1.2.1
cached-property==1.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
fitfile==1.1.9
Flask==3.0.0
garmin-fit-sdk==21.158.0
garmindb==3.6.3
garth==0.5.2
greenlet==3.1.1
h11==0.14.0
idbutils==1.1.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
loguru==0.7.3
MarkupSafe==2.1.3
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.1
playwright==1.51.0
pluggy==1.5.0
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
pyee==12.1.1
pyproject_hooks==1.1.0
PySocks==1.7.1
pytest==8.3.5
pytest-base-url==2.1.0
pytest-playwright==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-for-android==2024.1.21
python-slugify==8.0.4
requests==2.32.3
requests-oauthlib==2.0.0
screeninfo==0.8.1
selenium==4.31.0
setuptools==72.2.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
tcxfile==1.0.4
text-unidecode==1.3
toml==0.10.2
tornado==6.4.2
tqdm==4.66.5
trio==0.29.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.0.1
win32_setctime==1.2.0
wsproto==1.2.0
